Fix bug in notification and added ViewTaskActivity

+++ app/src/main/java/com/devop/tasker/models/Task.java
@@ -105,7 +105,7 @@ public class Task implements Serializable {
                         cursor.getString(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_TITLE)),
                         cursor.getString(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_DESCRIPTION)),
                         cursor.getInt(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_IMPORTANCE)),
-                        cursor.getInt(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_DUE_TIME)));
+                        cursor.getLong(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_DUE_TIME)));
 
                 task.id = cursor.getInt(cursor.getColumnIndex(TaskColumns._ID));
                 task.status = cursor.getInt(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_STATUS));
@@ -131,7 +131,7 @@ public class Task implements Serializable {
                     cursor.getString(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_TITLE)),
                     cursor.getString(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_DESCRIPTION)),
                     cursor.getInt(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_IMPORTANCE)),
-                    cursor.getInt(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_DUE_TIME)));
+                    cursor.getLong(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_DUE_TIME)));
 
             task.id = cursor.getInt(cursor.getColumnIndex(TaskColumns._ID));
             task.status = cursor.getInt(cursor.getColumnIndex(TaskColumns.COLUMN_NAME_STATUS));
